requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=9.0.0
albumentations>=1.3.0
tqdm>=4.62.0
PyYAML>=6.0
matplotlib>=3.5.0
tensorboard>=2.10.0
pytorch-msssim>=1.0.0
scikit-learn>=1.0.0
scipy>=1.7.0
